LayoutTests/js/dom/Promise-resolve-with-itself-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS rejected
PASS result is "TypeError: Cannot resolve a promise with itself"
PASS successfullyParsed is true

TEST COMPLETE
